Move `api` to `persistence/remote` and add standard stream test-suite.
