+++ internal/grpcx/doc.go
@@ -1,0 +1,2 @@
+// Package grpcx contains utilities for writing gRPC clients and servers.
+package grpcx
