+++ persistence/remote/doc.go
@@ -1,0 +1,3 @@
+// Package remote contains a persistence.Stream implementation that streams
+// messages via the messaging API.
+package remote
