+++ persistence/remote/ginkgo_test.go
@@ -1,4 +1,4 @@
-package api_test
+package remote_test
 
 import (
 	"reflect"
