+++ persistence/remote/stream.go
@@ -1,4 +1,4 @@
-package api
+package remote
 
 import (
 	"context"
